Move graphql to src/graphql Refactor some func
