+++ src/App.ts
@@ -10,9 +10,9 @@ import cors from "cors"
 import { ApolloServer } from 'apollo-server-express'
 import depthLimit from 'graphql-depth-limit'
 import { importSchema } from 'graphql-import';
-import resolvers from './resolvers'
-import { ReqContext } from './types/server.types';
-import { peopleFromHouseholdLoader, householdLoader } from './loader'
+import resolvers from './graphql/resolvers'
+import { ReqContext } from './graphql/types/server.types';
+import { peopleFromHouseholdLoader, householdLoader } from './graphql/loader'
 import { validateToken } from './helpers/account';
 
 export const init = async () => {
@@ -40,7 +40,7 @@ export const init = async () => {
 
     const server = app.setConfig(configFunction).build();
     const graphQLServer = new ApolloServer({
-      typeDefs: importSchema('src/schema/schema.graphql'),
+      typeDefs: importSchema('src/graphql/schema/schema.graphql'),
       resolvers,
       validationRules: [depthLimit(5)],
       context: (ctx: ReqContext) => {
