+++ src/graphql/types/server.types.ts
@@ -1,5 +1,5 @@
 import { ExpressContext } from 'apollo-server-express';
-import { IMe } from '../helpers/account';
+import { IMe } from '../../helpers/account';
 import { HouseholdLoader, PeopleFromHouseholdLoader } from '../loader';
 
 export type ReqContext = ExpressContext & {
